requirements.txt
python-dotenv==1.1.0
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
drf-yasg==1.21.10
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
drf-yasg==1.21.10
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
drf-yasg==1.21.10
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
